OUR-388 Search for packages, images are to big in list
http://issues.umbraco.org/issue/OUR-388

In the initial project listing, the images are set to
"?bgcolor=fff&width=50&height=50&format=png" but in the search listing a
little bit further up in the same file, the images are not restricted to
any size. That made Internet Explorer and Firefox go bananas, but for
some reasion not Chrome.
Setting the same queries on the search listing fixes this.